requirements.txt
openai==1.43.0
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.38.0
pycryptodome==3.15.0
